@@ -1,6 +1,6 @@
 package jt.kundream.adapter.rviewadapter.base;
 
-import android.support.v4.util.SparseArrayCompat;
+import androidx.collection.SparseArrayCompat;
 
 
 /**
