@@ -10,7 +10,7 @@ package jt.kundream.utils;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.WindowManager;
 
 /**
