@@ -2,10 +2,10 @@ package jt.kundream.view;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v4.BuildConfig;
-import android.support.v4.view.ViewCompat;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import jt.kundream.BuildConfig;;
+import androidx.core.view.ViewCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.View;
 
