@@ -22,15 +22,15 @@ public class DaoMaster extends AbstractDaoMaster {
     /** Creates underlying database table using DAOs. */
     public static void createAllTables(Database db, boolean ifNotExists) {
         GoodsBeanDaoDao.createTable(db, ifNotExists);
-        OrderBeanDaoDao.createTable(db, ifNotExists);
         OrderIdDaoDao.createTable(db, ifNotExists);
+        OrderBeanDaoDao.createTable(db, ifNotExists);
     }
 
     /** Drops underlying database table using DAOs. */
     public static void dropAllTables(Database db, boolean ifExists) {
         GoodsBeanDaoDao.dropTable(db, ifExists);
-        OrderBeanDaoDao.dropTable(db, ifExists);
         OrderIdDaoDao.dropTable(db, ifExists);
+        OrderBeanDaoDao.dropTable(db, ifExists);
     }
 
     /**
@@ -50,8 +50,8 @@ public class DaoMaster extends AbstractDaoMaster {
     public DaoMaster(Database db) {
         super(db, SCHEMA_VERSION);
         registerDaoClass(GoodsBeanDaoDao.class);
-        registerDaoClass(OrderBeanDaoDao.class);
         registerDaoClass(OrderIdDaoDao.class);
+        registerDaoClass(OrderBeanDaoDao.class);
     }
 
     public DaoSession newSession() {
