@@ -9,12 +9,12 @@ import org.greenrobot.greendao.identityscope.IdentityScopeType;
 import org.greenrobot.greendao.internal.DaoConfig;
 
 import cn.iyunbei.handself.bean.GoodsBeanDao;
-import cn.iyunbei.handself.bean.OrderBeanDao;
 import cn.iyunbei.handself.bean.OrderIdDao;
+import cn.iyunbei.handself.bean.OrderBeanDao;
 
 import cn.iyunbei.handself.greendao.GoodsBeanDaoDao;
-import cn.iyunbei.handself.greendao.OrderBeanDaoDao;
 import cn.iyunbei.handself.greendao.OrderIdDaoDao;
+import cn.iyunbei.handself.greendao.OrderBeanDaoDao;
 
 // THIS CODE IS GENERATED BY greenDAO, DO NOT EDIT.
 
@@ -26,12 +26,12 @@ import cn.iyunbei.handself.greendao.OrderIdDaoDao;
 public class DaoSession extends AbstractDaoSession {
 
     private final DaoConfig goodsBeanDaoDaoConfig;
-    private final DaoConfig orderBeanDaoDaoConfig;
     private final DaoConfig orderIdDaoDaoConfig;
+    private final DaoConfig orderBeanDaoDaoConfig;
 
     private final GoodsBeanDaoDao goodsBeanDaoDao;
-    private final OrderBeanDaoDao orderBeanDaoDao;
     private final OrderIdDaoDao orderIdDaoDao;
+    private final OrderBeanDaoDao orderBeanDaoDao;
 
     public DaoSession(Database db, IdentityScopeType type, Map<Class<? extends AbstractDao<?, ?>>, DaoConfig>
             daoConfigMap) {
@@ -40,37 +40,37 @@ public class DaoSession extends AbstractDaoSession {
         goodsBeanDaoDaoConfig = daoConfigMap.get(GoodsBeanDaoDao.class).clone();
         goodsBeanDaoDaoConfig.initIdentityScope(type);
 
-        orderBeanDaoDaoConfig = daoConfigMap.get(OrderBeanDaoDao.class).clone();
-        orderBeanDaoDaoConfig.initIdentityScope(type);
-
         orderIdDaoDaoConfig = daoConfigMap.get(OrderIdDaoDao.class).clone();
         orderIdDaoDaoConfig.initIdentityScope(type);
 
+        orderBeanDaoDaoConfig = daoConfigMap.get(OrderBeanDaoDao.class).clone();
+        orderBeanDaoDaoConfig.initIdentityScope(type);
+
         goodsBeanDaoDao = new GoodsBeanDaoDao(goodsBeanDaoDaoConfig, this);
-        orderBeanDaoDao = new OrderBeanDaoDao(orderBeanDaoDaoConfig, this);
         orderIdDaoDao = new OrderIdDaoDao(orderIdDaoDaoConfig, this);
+        orderBeanDaoDao = new OrderBeanDaoDao(orderBeanDaoDaoConfig, this);
 
         registerDao(GoodsBeanDao.class, goodsBeanDaoDao);
-        registerDao(OrderBeanDao.class, orderBeanDaoDao);
         registerDao(OrderIdDao.class, orderIdDaoDao);
+        registerDao(OrderBeanDao.class, orderBeanDaoDao);
     }
     
     public void clear() {
         goodsBeanDaoDaoConfig.clearIdentityScope();
-        orderBeanDaoDaoConfig.clearIdentityScope();
         orderIdDaoDaoConfig.clearIdentityScope();
+        orderBeanDaoDaoConfig.clearIdentityScope();
     }
 
     public GoodsBeanDaoDao getGoodsBeanDaoDao() {
         return goodsBeanDaoDao;
     }
 
-    public OrderBeanDaoDao getOrderBeanDaoDao() {
-        return orderBeanDaoDao;
-    }
-
     public OrderIdDaoDao getOrderIdDaoDao() {
         return orderIdDaoDao;
+    }
+
+    public OrderBeanDaoDao getOrderBeanDaoDao() {
+        return orderBeanDaoDao;
     }
 
 }
